[refactor] Co-loacte non-interface related functions
This makes the interface.go file more purposeful and also ties similar logic
together in goldie.go
