+++ interface.go
@@ -1,37 +1,13 @@
 package goldie
 
 import (
-	"flag"
-	"fmt"
 	"os"
 	"testing"
-
-	"github.com/pmezard/go-difflib/difflib"
-	"github.com/sergi/go-diff/diffmatchpatch"
 )
 
-const (
-	// FixtureDir is the folder name for where the fixtures are stored. It's
-	// relative to the "go test" path.
-	defaultFixtureDir = "testdata"
-
-	// FileNameSuffix is the suffix appended to the fixtures. Set to empty
-	// string to disable file name suffixes.
-	defaultFileNameSuffix = ".golden"
-
-	// FilePerms is used to set the permissions on the golden fixture files.
-	defaultFilePerms os.FileMode = 0644
-
-	// DirPerms is used to set the permissions on the golden fixture folder.
-	defaultDirPerms os.FileMode = 0755
-)
-
-var (
-	// update determines if the actual received data should be written to the
-	// golden files or not. This should be true when you need to update the
-	// data, but false when actually running the tests.
-	update = flag.Bool("update", false, "Update golden test file fixture")
-)
+// Compile time assurance
+var _ Tester = (*goldie)(nil)
+var _ OptionProcessor = (*goldie)(nil)
 
 // Option defines the signature of a functional option method that can apply
 // options to an OptionProcessor.
@@ -171,53 +147,4 @@ func WithSubTestNameForDir(use bool) Option {
 	return func(o OptionProcessor) error {
 		return o.WithSubTestNameForDir(use)
 	}
-}
-
-// === Create new testers ==================================
-
-// New creates a new golden file tester. If there is an issue with applying any
-// of the options, an error will be reported and t.FailNow() will be called.
-func New(t *testing.T, options ...Option) *goldie {
-	g := goldie{
-		fixtureDir:     defaultFixtureDir,
-		fileNameSuffix: defaultFileNameSuffix,
-		filePerms:      defaultFilePerms,
-		dirPerms:       defaultDirPerms,
-	}
-
-	var err error
-	for _, option := range options {
-		err = option(&g)
-		if err != nil {
-			t.Error(fmt.Errorf("Could not apply option: %w", err))
-			t.FailNow()
-		}
-	}
-
-	return &g
-}
-
-// Diff generates a string that shows the difference between the actual and the
-// expected. This method could be called in your own DiffFn in case you want
-// to leverage any of the engines defined.
-func Diff(engine DiffEngine, actual string, expected string) string {
-	var diff string
-	switch engine {
-	case ClassicDiff:
-		diff, _ = difflib.GetUnifiedDiffString(difflib.UnifiedDiff{
-			A:        difflib.SplitLines(expected),
-			B:        difflib.SplitLines(actual),
-			FromFile: "Expected",
-			FromDate: "",
-			ToFile:   "Actual",
-			ToDate:   "",
-			Context:  1,
-		})
-
-	case ColoredDiff:
-		dmp := diffmatchpatch.New()
-		diffs := dmp.DiffMain(actual, expected, false)
-		diff = dmp.DiffPrettyText(diffs)
-	}
-	return diff
 }
